Persistance

+++ src/main/java/com/company/CourseDao.java
@@ -5,8 +5,8 @@ import java.util.List;
 public interface CourseDao {
 
     public void store(Course course);
-    public void delete(Long courseId);
-    public Course findById(Long courseId);
+    public void delete(Integer courseId);
+    public Course findById(Integer courseId);
     public List<Course> findAll();
 
 }
